+++ Project5/db/migrate/20200501190912_add_email_to_login_request.rb
@@ -1,0 +1,5 @@
+class AddEmailToLoginRequest < ActiveRecord::Migration[6.0]
+  def change
+    add_column :login_requests, :email, :string
+  end
+end