+++ Project5/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_05_01_181454) do
+ActiveRecord::Schema.define(version: 2020_05_01_190912) do
 
   create_table "applications", force: :cascade do |t|
     t.string "course_number"
@@ -66,6 +66,7 @@ ActiveRecord::Schema.define(version: 2020_05_01_181454) do
     t.string "department"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+    t.string "email"
   end
 
   create_table "recommendations", force: :cascade do |t|